Extract Executors to indie modules
